+++ src/download_thread.cc
@@ -132,8 +132,8 @@ RVal do_foreground_execute(C *inst,
   return return_box.take();
 }
 
-void background_status::Fetched(unsigned long Size,
-				unsigned long ResumePoint)
+void background_status::Fetched(unsigned long long Size,
+				unsigned long long ResumePoint)
 {
   do_foreground_execute(real_status, Size, ResumePoint,
 			&download_signal_log::Fetched,
