+++ src/generic/apt/download_queue.h
@@ -71,8 +71,8 @@ namespace aptitude
      *  The default implementation does nothing.
      */
     virtual void partial_download(const temp::name &filename,
-				  unsigned long currentSize,
-				  unsigned long totalSize)
+				  unsigned long long currentSize,
+				  unsigned long long totalSize)
     {
     }
