+++ src/generic/views/mocks/download_progress.h
@@ -46,14 +46,14 @@ namespace aptitude
         MOCK_METHOD1(update_progress, bool(const status &));
         MOCK_METHOD3(file_started, void(const std::string &,
                                         const boost::optional<unsigned long> &,
-                                        const boost::optional<unsigned long> &));
+                                        const boost::optional<unsigned long long> &));
         MOCK_METHOD4(error, void(bool,
                                  const std::string &,
                                  const std::string &,
                                  const boost::optional<unsigned long> &));
         MOCK_METHOD2(file_finished, void(const std::string &,
                                          const boost::optional<unsigned long> &));
-        MOCK_METHOD3(done, void(double, unsigned long, double));
+        MOCK_METHOD3(done, void(unsigned long long, unsigned long long, unsigned long long));
         MOCK_METHOD3(media_change, void(const std::string &,
                                         const std::string &,
                                         const sigc::slot1<void, bool> &));
